requirements.txt
fastapi
uvicorn
pillow
pypdf
python-multipart
pdf2docx
python-docx
jinja2
PyMuPDF
reportlab
weasyprint